+++ src/CMakeLists.txt
@@ -99,6 +99,7 @@ add_library(parthenon
   bvals/cc/bvals_utils.hpp
   bvals/cc/build_boundary_buffers.cpp
   bvals/cc/bnd_info.cpp
+  bvals/cc/bnd_info.hpp
   bvals/cc/bvals_in_one.cpp
   bvals/cc/flux_correction.cpp 
   bvals/cc/tag_map.cpp
