+++ src/bvals/bvals.hpp
@@ -27,6 +27,7 @@
 #include "parthenon_mpi.hpp"
 
 #include "bvals/bvals_interfaces.hpp"
+#include "bvals/cc/bnd_info.hpp"
 #include "bvals/cc/bvals_cc_in_one.hpp"
 #include "defs.hpp"
 #include "mesh/domain.hpp"
