+++ src/bvals/cc/bnd_info.cpp
@@ -22,8 +22,8 @@
 #include <vector>
 
 #include "basic_types.hpp"
+#include "bvals/cc/bnd_info.hpp"
 #include "bvals/bvals_interfaces.hpp"
-#include "bvals_cc_in_one.hpp"
 #include "config.hpp"
 #include "globals.hpp"
 #include "interface/variable.hpp"
