+++ src/bvals/cc/bvals_cc_in_one.hpp
@@ -38,19 +38,6 @@ template <typename T>
 class CellVariable;
 
 namespace cell_centered_bvars {
-void CalcIndicesSetSame(int ox, int &s, int &e, const IndexRange &bounds);
-void CalcIndicesSetFromCoarser(const int &ox, int &s, int &e, const IndexRange &bounds,
-                               const std::int64_t &lx, const int &cng, bool include_dim);
-void CalcIndicesSetFromFiner(int &si, int &ei, int &sj, int &ej, int &sk, int &ek,
-                             const NeighborBlock &nb, MeshBlock *pmb);
-void CalcIndicesLoadSame(int ox, int &s, int &e, const IndexRange &bounds);
-void CalcIndicesLoadToFiner(int &si, int &ei, int &sj, int &ej, int &sk, int &ek,
-                            const NeighborBlock &nb, MeshBlock *pmb);
-
-int GetBufferSize(std::shared_ptr<MeshBlock> pmb, const NeighborBlock &nb,
-                  std::shared_ptr<CellVariable<Real>> v);
-
-TaskStatus BuildBoundaryBuffers(std::shared_ptr<MeshData<Real>> &md);
 
 template <BoundaryType bound_type>
 TaskStatus SendBoundBufs(std::shared_ptr<MeshData<Real>> &md);
@@ -79,84 +66,8 @@ TaskStatus LoadAndSendFluxCorrections(std::shared_ptr<MeshData<Real>> &md);
 TaskStatus ReceiveFluxCorrections(std::shared_ptr<MeshData<Real>> &md);
 TaskStatus SetFluxCorrections(std::shared_ptr<MeshData<Real>> &md);
 
-struct BndInfo {
-  int si = 0;
-  int ei = 0;
-  int sj = 0;
-  int ej = 0;
-  int sk = 0;
-  int ek = 0;
-
-  int Nt = 0;
-  int Nu = 0;
-  int Nv = 0;
-
-  CoordinateDirection dir;
-  bool allocated = true;
-  RefinementOp_t refinement_op = RefinementOp_t::None;
-  Coordinates_t coords, coarse_coords; // coords
-
-  buf_pool_t<Real>::weak_t buf; // comm buffer from pool
-  ParArray6D<Real> var;         // data variable used for comms
-  ParArray6D<Real> fine;        // fine data variable for prolongation/restriction
-  ParArray6D<Real> coarse;      // coarse data variable for prolongation/restriction
-
-  static BndInfo GetSendBndInfo(std::shared_ptr<MeshBlock> pmb, const NeighborBlock &nb,
-                                std::shared_ptr<CellVariable<Real>> v,
-                                CommBuffer<buf_pool_t<Real>::owner_t> *buf);
-  static BndInfo GetSetBndInfo(std::shared_ptr<MeshBlock> pmb, const NeighborBlock &nb,
-                               std::shared_ptr<CellVariable<Real>> v,
-                               CommBuffer<buf_pool_t<Real>::owner_t> *buf);
-  static BndInfo GetSendCCFluxCor(std::shared_ptr<MeshBlock> pmb, const NeighborBlock &nb,
-                                  std::shared_ptr<CellVariable<Real>> v,
-                                  CommBuffer<buf_pool_t<Real>::owner_t> *buf);
-  static BndInfo GetSetCCFluxCor(std::shared_ptr<MeshBlock> pmb, const NeighborBlock &nb,
-                                 std::shared_ptr<CellVariable<Real>> v,
-                                 CommBuffer<buf_pool_t<Real>::owner_t> *buf);
-};
-
-int GetBufferSize(std::shared_ptr<MeshBlock> pmb, const NeighborBlock &nb,
-                  std::shared_ptr<CellVariable<Real>> v);
-
-using BufferCache_t = ParArray1D<BndInfo>;
-using BufferCacheHost_t = typename BufferCache_t::HostMirror;
-
-// This is just a struct to cleanly hold all of the information it is useful to cache
-// for the block boundary communication routines. A copy of it is contained in MeshData.
-struct BvarsSubCache_t {
-  void clear() {
-    buf_vec.clear();
-    idx_vec.clear();
-    if (sending_non_zero_flags.KokkosView().is_allocated())
-      sending_non_zero_flags = ParArray1D<bool>{};
-    if (sending_non_zero_flags_h.KokkosView().is_allocated())
-      sending_non_zero_flags_h = ParArray1D<bool>::host_mirror_type{};
-    bnd_info = BufferCache_t{};
-    bnd_info_h = BufferCache_t::host_mirror_type{};
-  }
-
-  std::vector<std::size_t> idx_vec;
-  std::vector<CommBuffer<buf_pool_t<Real>::owner_t> *> buf_vec;
-  ParArray1D<bool> sending_non_zero_flags;
-  ParArray1D<bool>::host_mirror_type sending_non_zero_flags_h;
-
-  BufferCache_t bnd_info{};
-  BufferCache_t::host_mirror_type bnd_info_h{};
-};
-
-struct BvarsCache_t {
-  // The five here corresponds to the current size of the BoundaryType enum
-  std::array<BvarsSubCache_t, 5 * 2> caches;
-  auto &GetSubCache(BoundaryType boundType, bool send) {
-    return caches[2 * static_cast<int>(boundType) + send];
-  }
-  // auto &operator[](BoundaryType boundType) { return
-  // caches[static_cast<int>(boundType)]; }
-  void clear() {
-    for (int i = 0; i < caches.size(); ++i)
-      caches[i].clear();
-  }
-};
+// This task should not be called in down stream code 
+TaskStatus BuildBoundaryBuffers(std::shared_ptr<MeshData<Real>> &md);
 
 } // namespace cell_centered_bvars
 } // namespace parthenon
