+++ src/bvals/cc/bvals_in_one.cpp
@@ -22,6 +22,7 @@
 #include <string>
 #include <vector>
 
+#include "bnd_info.hpp"
 #include "bvals_cc_in_one.hpp"
 #include "bvals_utils.hpp"
 #include "config.hpp"
