+++ src/bvals/cc/flux_correction.cpp
@@ -22,6 +22,7 @@
 #include <type_traits>
 #include <vector>
 
+#include "bnd_info.hpp"
 #include "bvals_cc_in_one.hpp"
 #include "bvals_utils.hpp"
 #include "config.hpp"
