+++ src/interface/mesh_data.hpp
@@ -22,7 +22,7 @@
 #include <utility>
 #include <vector>
 
-#include "bvals/cc/bvals_cc_in_one.hpp"
+#include "bvals/cc/bnd_info.hpp"
 #include "interface/sparse_pack_base.hpp"
 #include "interface/variable_pack.hpp"
 #include "mesh/domain.hpp"
